+++ CHANGES.md
@@ -1,7 +1,9 @@
+* 0.3.4 - 2018-04-06
+  * Config: take a `launcher`
 * 0.3.3 - 2018-04-06
-  * Take a `borderColor` in config
+  * Config: take a `borderColor`
 * 0.3.2 - 2018-04-06
-  * Take a `borderWidth` in config
+  * Config: take a `borderWidth`
 * 0.3.0 - 2018-04-05
   * Taking a config file now works
   * `startupPrograms` in the config file now works